requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
openai>=1.12.0
groq>=0.2.0
tavily-python>=0.2.0
boto3==1.29.7
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.26.2
aiofiles==23.2.1
email-validator>=2.0.0

